* fixed compiler warnings * fixed code analyst warnings

+++ src/game/g_local.h
@@ -175,10 +175,10 @@ extern game_export_t globals;
 
 extern edict_t *g_edicts;
 
-#define	FOFS(x) (int)&(((edict_t *)0)->x)
-#define	STOFS(x) (int)&(((spawn_temp_t *)0)->x)
-#define	LLOFS(x) (int)&(((level_locals_t *)0)->x)
-#define	CLOFS(x) (int)&(((player_t *)0)->x)
+#define	FOFS(x) (size_t)&(((edict_t *)0)->x)
+#define	STOFS(x) (size_t)&(((spawn_temp_t *)0)->x)
+#define	LLOFS(x) (size_t)&(((level_locals_t *)0)->x)
+#define	CLOFS(x) (size_t)&(((player_t *)0)->x)
 
 #define random()	((rand () & 0x7fff) / ((float)0x7fff))
 #define crandom()	(2.0 * (random() - 0.5))
@@ -286,7 +286,7 @@ typedef enum {
 
 typedef struct {
 	char *name;
-	int ofs;
+	size_t ofs;
 	fieldtype_t type;
 	int flags;
 } field_t;
